My applyColorMap for visualize details in image
This is an extended version of applyColorMap from OpenCV, so based on applyColorMap function.
Why it is developed: to see the change of some pixel (homogenous ) in detail. 
The image will be changed to heatmap according to your own settings. 
How to use (example in file): you need to set the image, min value, max value. the option is the method (style) of the heatmap. 
Enjoy it!